refactor: ajuste em nome do pacote e estrutura de pasta
