+++ app/application/usecase/account/create_account.go
@@ -1,4 +1,4 @@
-package usecase
+package account
 
 import (
 	"context"
