+++ app/application/usecase/account/errors.go
@@ -1,4 +1,4 @@
-package usecase
+package account
 
 import "errors"
