+++ app/application/usecase/account/get_all_accounts.go
@@ -1,4 +1,4 @@
-package usecase
+package account
 
 import (
 	"context"
