+++ app/application/usecase/account/get_balance.go
@@ -1,4 +1,4 @@
-package usecase
+package account
 
 import (
 	"context"
