+++ app/application/usecase/account/usecase.go
@@ -1,4 +1,4 @@
-package usecase
+package account
 
 import (
 	"context"
