--- app/application/usecase/get_balance_test.go
@@ -1,1 +1,0 @@
-package usecase
